+++ tests/e2e/Adapter/MySQLTest.php
@@ -1,62 +1,62 @@
 <?php
-
-namespace Tests\E2E\Adapter;
-
-use PDO;
-use Redis;
-use Utopia\Cache\Adapter\Redis as RedisAdapter;
-use Utopia\Cache\Cache;
-use Utopia\Database\Adapter\MySQL;
-use Utopia\Database\Database;
-
-class MySQLTest extends Base
-{
-    public static ?Database $database = null;
-    protected static string $namespace;
-
-    // Remove once all methods are implemented
-    /**
-     * Return name of adapter
-     *
-     * @return string
-     */
-    public static function getAdapterName(): string
-    {
-        return "mysql";
-    }
-
-    /**
-     * @return Database
-     */
-    public static function getDatabase(): Database
-    {
-        if (!is_null(self::$database)) {
-            return self::$database;
-        }
-
-        $dbHost = 'mysql';
-        $dbPort = '3307';
-        $dbUser = 'root';
-        $dbPass = 'password';
-
-        $pdo = new PDO("mysql:host={$dbHost};port={$dbPort};charset=utf8mb4", $dbUser, $dbPass, MySQL::getPDOAttributes());
-
-        $redis = new Redis();
-        $redis->connect('redis', 6379);
-        $redis->flushAll();
-
-        $cache = new Cache(new RedisAdapter($redis));
-
-        $database = new Database(new MySQL($pdo), $cache);
-        $database->setDatabase('utopiaTests');
-        $database->setNamespace(static::$namespace = 'myapp_' . uniqid());
-
-        if ($database->exists('utopiaTests')) {
-            $database->delete('utopiaTests');
-        }
-
-        $database->create();
-
-        return self::$database = $database;
-    }
-}
+//
+//namespace Tests\E2E\Adapter;
+//
+//use PDO;
+//use Redis;
+//use Utopia\Cache\Adapter\Redis as RedisAdapter;
+//use Utopia\Cache\Cache;
+//use Utopia\Database\Adapter\MySQL;
+//use Utopia\Database\Database;
+//
+//class MySQLTest extends Base
+//{
+//    public static ?Database $database = null;
+//    protected static string $namespace;
+//
+//    // Remove once all methods are implemented
+//    /**
+//     * Return name of adapter
+//     *
+//     * @return string
+//     */
+//    public static function getAdapterName(): string
+//    {
+//        return "mysql";
+//    }
+//
+//    /**
+//     * @return Database
+//     */
+//    public static function getDatabase(): Database
+//    {
+//        if (!is_null(self::$database)) {
+//            return self::$database;
+//        }
+//
+//        $dbHost = 'mysql';
+//        $dbPort = '3307';
+//        $dbUser = 'root';
+//        $dbPass = 'password';
+//
+//        $pdo = new PDO("mysql:host={$dbHost};port={$dbPort};charset=utf8mb4", $dbUser, $dbPass, MySQL::getPDOAttributes());
+//
+//        $redis = new Redis();
+//        $redis->connect('redis', 6379);
+//        $redis->flushAll();
+//
+//        $cache = new Cache(new RedisAdapter($redis));
+//
+//        $database = new Database(new MySQL($pdo), $cache);
+//        $database->setDatabase('utopiaTests');
+//        $database->setNamespace(static::$namespace = 'myapp_' . uniqid());
+//
+//        if ($database->exists('utopiaTests')) {
+//            $database->delete('utopiaTests');
+//        }
+//
+//        $database->create();
+//
+//        return self::$database = $database;
+//    }
+//}
